requirements.txt
ccxt==4.5.28
pytest==7.4.3
python-dotenv==1.2.1
tabulate==0.9.0
